# refactor(s3stream/object-wal): complete appends sequentially