--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
@@ -200,6 +200,7 @@ public class ObjectWALService implements WriteAheadLog {
 
         public RecoverIterator(List<RecordAccumulator.WALObject> objectList, ObjectStorage objectStorage,
             int readAheadObjectSize) {
+            // TODO: check and drop discontinuous objects.
             this.objectList = objectList;
             this.objectStorage = objectStorage;
             this.readAheadObjectSize = readAheadObjectSize;