--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
@@ -77,7 +77,7 @@ public class ObjectWALService implements WriteAheadLog {
     }
 
     // Visible for testing.
-    protected RecordAccumulator accumulator() {
+    RecordAccumulator accumulator() {
         return accumulator;
     }