--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/ObjectWALService.java
@@ -245,7 +245,7 @@ public class ObjectWALService implements WriteAheadLog {
 
             // Check header
             WALObjectHeader header = WALObjectHeader.unmarshal(dataBuffer);
-            dataBuffer.skipBytes(WALObjectHeader.WAL_HEADER_SIZE_V0);
+            dataBuffer.skipBytes(header.size());
 
             if (skipStickyRecord && header.stickyRecordLength() != 0) {
                 dataBuffer.skipBytes((int) header.stickyRecordLength());