--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/RecordAccumulator.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/RecordAccumulator.java
@@ -523,6 +523,7 @@ public class RecordAccumulator implements Closeable {
                     objectMap.put(endOffset, new WALObject(result.bucket(), path, firstOffset, objectLength));
                     objectDataBytes.addAndGet(objectLength);
 
+                    // TODO: only remove records when all upload tasks before `firstOffset` is finished
                     List<Record> uploadedRecords = uploadMap.remove(firstOffset);
 
                     // Update flushed offset