--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/RecordAccumulator.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/RecordAccumulator.java
@@ -42,6 +42,7 @@ import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.PriorityQueue;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
@@ -68,7 +69,7 @@ public class RecordAccumulator implements Closeable {
     protected final Time time;
     protected final ObjectStorage objectStorage;
     private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
-    private final ConcurrentNavigableMap<Long /* inclusive first offset */, List<Record>> uploadMap = new ConcurrentSkipListMap<>();
+    private final ConcurrentNavigableMap<Long /* inclusive first offset */, UploadTask> uploadMap = new ConcurrentSkipListMap<>();
     private final ConcurrentNavigableMap<Pair<Long /* epoch */, Long /* exclusive end offset */>, WALObject> previousObjectMap = new ConcurrentSkipListMap<>();
     private final ConcurrentNavigableMap<Long /* exclusive end offset */, WALObject> objectMap = new ConcurrentSkipListMap<>();
     private final String nodePrefix;
@@ -504,7 +505,7 @@ public class RecordAccumulator implements Closeable {
 
         // Trigger upload.
         int objectLength = objectBuffer.readableBytes();
-        uploadMap.put(firstOffset, recordList);
+        uploadMap.put(firstOffset, new UploadTask(recordList));
 
         // Enable fast retry.
         ObjectStorage.WriteOptions writeOptions = new ObjectStorage.WriteOptions().enableFastRetry(true);
@@ -523,8 +524,13 @@ public class RecordAccumulator implements Closeable {
                     objectMap.put(endOffset, new WALObject(result.bucket(), path, firstOffset, objectLength));
                     objectDataBytes.addAndGet(objectLength);
 
-                    // TODO: only remove records when all upload tasks before `firstOffset` is finished
-                    List<Record> uploadedRecords = uploadMap.remove(firstOffset);
+                    uploadMap.get(firstOffset).markFinished();
+                    List<UploadTask> finishedTasks = new ArrayList<>();
+                    // Remove consecutive completed tasks from head.
+                    Map.Entry<Long, UploadTask> entry;
+                    while ((entry = uploadMap.firstEntry()) != null && entry.getValue().isFinished()) {
+                        finishedTasks.add(uploadMap.remove(entry.getKey()));
+                    }
 
                     // Update flushed offset
                     if (!uploadMap.isEmpty()) {
@@ -536,7 +542,11 @@ public class RecordAccumulator implements Closeable {
                     }
 
                     // Release lock and complete future in callback thread.
-                    callbackService.submit(() -> uploadedRecords.forEach(record -> record.future.complete(flushedOffset::get)));
+                    for (UploadTask task : finishedTasks) {
+                        callbackService.submit(() ->
+                            task.records().forEach(record -> record.future.complete(flushedOffset::get))
+                        );
+                    }
                 } finally {
                     lock.writeLock().unlock();
                 }
@@ -545,7 +555,7 @@ public class RecordAccumulator implements Closeable {
                 bufferedDataBytes.addAndGet(-dataLength);
                 throwable = ExceptionUtils.getRootCause(throwable);
                 if (throwable instanceof WALFencedException) {
-                    List<Record> uploadedRecords = uploadMap.remove(firstOffset);
+                    List<Record> uploadedRecords = uploadMap.remove(firstOffset).records();
                     Throwable finalThrowable = throwable;
                     // Release lock and complete future in callback thread.
                     callbackService.submit(() -> uploadedRecords.forEach(record -> record.future.completeExceptionally(finalThrowable)));
@@ -567,6 +577,27 @@ public class RecordAccumulator implements Closeable {
             ObjectWALMetricsManager.recordOperationLatency(time.nanoseconds() - startTime, "upload", throwable == null);
             ObjectWALMetricsManager.recordOperationDataSize(objectLength, "upload");
         });
+    }
+
+    private static class UploadTask {
+        private final List<Record> records;
+        private boolean finished = false;
+
+        public UploadTask(List<Record> records) {
+            this.records = records;
+        }
+
+        public List<Record> records() {
+            return records;
+        }
+
+        public void markFinished() {
+            this.finished = true;
+        }
+
+        public boolean isFinished() {
+            return finished;
+        }
     }
 
     protected static class Record {