… s3stream/src/main/java/com/automq/stream/s3/wal/impl/object/WALObjectHeader.java
… s3stream/src/test/java/com/automq/stream/s3/wal/impl/object/MockObjectStorage.java
… s3stream/src/test/java/com/automq/stream/s3/wal/impl/object/ObjectWALServiceTest.java
… s3stream/src/test/java/com/automq/stream/s3/wal/impl/object/RecordAccumulatorTest.java
… s3stream/src/test/java/com/automq/stream/s3/wal/impl/object/WALObjectHeaderTest.java