@daniel-y 2025-04-24
Should add unit tests to cover this refactor, and consider creating subsequent issues and PRs to simplify the upper-level logic.